# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/subscriptions": {
            "get": {
                "description": "Пагинированный список подписок",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Получение списка подписок",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Максимум записей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/T_internal_entity.Subscription"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создание новой подписки для пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Создать подписку",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_v01.CreateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданной подписки",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/subscriptions/total": {
            "post": {
                "description": "Фильтрация по пользователю, сервису и периоду (через JSON)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Получение общей стоимости подписок",
                "parameters": [
                    {
                        "description": "Фильтры",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_v01.TotalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/subscriptions/{id}": {
            "get": {
                "description": "Возвращает подписку по UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Получить подписку по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID подписки (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/T_internal_entity.Subscription"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновление полей существующей подписки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Обновить подписку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID подписки (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_http_v01.UpdateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет подписку по её UUID (мягкое или жёсткое удаление зависит от реализации)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Удалить подписку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID подписки (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/T_internal_api_http_v01_errors.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "T_internal_api_http_v01_errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "\"description\"": "\"must be valid UUID\"}",
                        "{\"field\"": "\"user_id\""
                    }
                },
                "message": {
                    "type": "string",
                    "example": "invalid request"
                },
                "reason": {
                    "type": "string",
                    "example": "invalid_guid"
                }
            }
        },
        "T_internal_entity.Subscription": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Для информации",
                    "type": "string"
                },
                "end_date": {
                    "description": "Конец подписки (опционально)",
                    "type": "string"
                },
                "id": {
                    "description": "Уникальный идентификатор записи",
                    "type": "string"
                },
                "price": {
                    "description": "Цена в рублях",
                    "type": "integer"
                },
                "service_name": {
                    "description": "Название сервиса",
                    "type": "string"
                },
                "start_date": {
                    "description": "Начало подписки (месяц и год)",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Для информации",
                    "type": "string"
                },
                "user_id": {
                    "description": "UUID пользователя",
                    "type": "string"
                }
            }
        },
        "internal_api_http_v01.CreateSubscriptionRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "description": "можно пустым",
                    "type": "string",
                    "example": "12-2025"
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "description": "формат MM-YYYY",
                    "type": "string",
                    "example": "07-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "internal_api_http_v01.TotalRequest": {
            "type": "object",
            "properties": {
                "from": {
                    "description": "MM-YYYY",
                    "type": "string",
                    "example": "06-2024"
                },
                "service_name": {
                    "type": "string",
                    "example": "Netflix"
                },
                "to": {
                    "type": "string",
                    "example": "08-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        },
        "internal_api_http_v01.UpdateSubscriptionRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "description": "можно пустым",
                    "type": "string",
                    "example": "12-2025"
                },
                "price": {
                    "type": "integer",
                    "example": 400
                },
                "service_name": {
                    "type": "string",
                    "example": "Yandex Plus"
                },
                "start_date": {
                    "description": "формат MM-YYYY",
                    "type": "string",
                    "example": "07-2025"
                },
                "user_id": {
                    "type": "string",
                    "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
                }
            }
        }
    }
}